\name{BNDataGenerator-package}
\alias{BNDataGenerator-package}
\alias{BNDataGenerator}
\docType{package}

\title{
	Data Generator based on Bayesian Network Model
}

\details{

	\tabular{ll}{
		Package: \tab BN_Data_Generator\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2014-12-28\cr
		License: \tab GPL (>=2)\cr
		}
}

\author{ Jae-seong Yoo <praster1@gmail.com> }

\references{
	Jae-seong Yoo, (2014), "A Study on Comparison of Bayesian Network Structure Learning Algorithms for Selecting Appropriate Models", M.S. thesis, Department of Statistics, Korea University, Seoul.
}

\keyword{ Bayesian Network }

\seealso{
	See demo(gen_asia) or demo(topology).
}

\examples{
}