\name{check_cardinality}
\alias{check_cardinality}

\title{
	A checker of needs of input_Prob follows cardinality
}

\description{
	A checker of needs of input_Prob follows cardinality.
}

\usage{
	check_cardinality(arcs_mat, nodename=NULL, cardinality=NULL)
}

\arguments{
  \item{arcs_mat}{ A matrix that determines the arcs. }
  \item{nodename}{ The names of each nodes. }
  \item{cardinality}{ The cardinality of each nodes. }
}

\author{ Jae-seong Yoo <praster1@gmail.com> }