\name{fromto_to_mat}
\alias{fromto_to_mat}

\title{
	Convert from 'fromto' to 'matrix'
}

\description{
	Convert from 'fromto' to 'matrix' that determines the arcs.
}

\usage{
	fromto_to_mat(fromto, nodename)
}

\arguments{
  \item{fromto}{ A matrix form structured 'fromto' that determines the arcs.}
  \item{nodename}{ The names of each nodes. }
}

\author{ Jae-seong Yoo <praster1@gmail.com> }
