\name{mat_to_fromto}
\alias{mat_to_fromto}

\title{
	Convert from 'matrix' to 'fromto'
}

\description{
	Convert from 'matrix' to 'fromto' that determines the arcs.
}

\usage{
	mat_to_fromto(arcs_mat)
}

\arguments{
	\item{arcs_mat}{ A matrix that determines the arcs that determines the arcs. }
}

\author{ Jae-seong Yoo <praster1@gmail.com> }