\name{big_letters}
\alias{big_letters}

\title{
	Letters built
}
\description{
	The 26 or more lower-case letters of the Roman alphabet;
}

\usage{
	big_letters(len)
}

\arguments{
	  \item{len}{ Get or set the length of letters. }
}

\author{ Jae-seong Yoo <praster1@gmail.com> }

\examples{
	big_letters(26)
	big_letters(50)
	big_letters(100)
	big_letters(999)
}

\keyword{ letters }
