\name{check_cardinalities}
\alias{check_cardinalities}

\title{
	A checker of needs of input_Prob follows cardinality
}

\description{
	A checker of needs of input_Prob follows cardinality.
}

\usage{
	check_cardinalities(arcs_mat, node_names = NULL, cardinalities = NULL)
}

\arguments{
  \item{arcs_mat}{ A matrix that determines the arcs. }
  \item{node_names}{ The names of each nodes. }
  \item{cardinalities}{ The cardinalities of each nodes. }
}

\details{
}

\value{
}

\references{
}

\author{ Jae-seong Yoo <praster1@gmail.com> }

\note{
}

\seealso{
}

\examples{
}

\keyword{ }
